+++ flank-scripts/src/main/kotlin/flank/scripts/release/hub/DeleteOldRelease.kt
@@ -14,6 +14,6 @@ class DeleteOldReleaseCommand : CliktCommand(name = "deleteOldRelease", help = "
     }
 }
 
-fun deleteOldRelease(tag: String) = "$DELETE_RELEASE_COMMAND \"$tag\"".runCommand()
+fun deleteOldRelease(tag: String) = "$DELETE_RELEASE_COMMAND $tag".runCommand()
 
 private const val DELETE_RELEASE_COMMAND = "hub release delete"