+++ release_notes.md
@@ -1,5 +1,6 @@
 ## next (unreleased)
-- [#952](https://github.com/Flank/flank/pull/952) Fix version printing on Flank release
+- [#937](https://github.com/Flank/flank/pull/968) ([sloox](https://github.com/Sloox))
+- [#952](https://github.com/Flank/flank/pull/952) Fix version printing on Flank release ([sloox](https://github.com/Sloox))
 - [#950](https://github.com/Flank/flank/pull/950) Fix crash when --legacy-junit-result set. ([adamfilipow92](https://github.com/adamfilipow92))
 - [#948](https://github.com/Flank/flank/pull/948) Increment retry tries and change sync tag for jfrogSync. ([piotradamczyk5](https://github.com/piotradamczyk5))
 - [#946](https://github.com/Flank/flank/pull/946) Added tests for flank scripts. ([piotradamczyk5](https://github.com/piotradamczyk5))